+++ modules/home/eza.nix
@@ -1,0 +1,14 @@
+# Eza is a ls replacement
+{
+  programs.eza = {
+    enable = true;
+    icons = "auto";
+
+    extraOptions = [
+      "--group-directories-first"
+      "--no-quotes"
+      "--git-ignore"
+      "--icons=always"
+    ];
+  };
+}